pyproject.toml
[tool.poetry]
name = "lol-match-scraper"
version = "0.1.0"
description = "A repository for scraping ranked match data using the RIOT API."
authors = ["Mark Bugden"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "riot_api", from = "src" },
    { include = "scraping", from = "src"},
    { include = "processing", from = "src"}
]

[tool.poetry.dependencies]
python = ">=3.12, <4.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
tqdm = "^4.67.1"
loguru = "^0.7.3"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
